backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "doclayout-yolo>=0.0.4",
    "fastapi>=0.128.0",
    "mineru>=2.7.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "python-pptx>=0.6.23",
    "openai>=1.61.0",
    "google-genai>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "supabase>=2.27.3",
    "tokenizers>=0.22.2",
    "torch>=2.10.0",
    "torchvision>=0.19.0",
    "transformers<5",
    "ultralytics>=8.4.11",
    "uvicorn>=0.40.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
